Correcoes pontuais nas classes existentes e criacao de endpoints

+++ src/main/java/com/henriquecosta/criacao_de_api/DTO/MunicipalityDTO.java
@@ -21,9 +21,9 @@ public class MunicipalityDTO{
     public String cdFonteRecurso;
     public String cdTipoSeguro;
     @JsonProperty("QtdCusteio")
-    public int qtdCusteio;
+    public Double qtdCusteio;
     @JsonProperty("VlCusteio")
-    public int vlCusteio;
+    public Double vlCusteio;
     @JsonProperty("Atividade")
     public String atividade;
     public String cdModalidade;
